# cmd/server/openapi.yaml
openapi: "3.1.0"
info:
  title: Bastio API
  description: AI Security Gateway API
  version: "0.1.0"
  license:
    name: FSL-1.1-ALv2
    url: https://fsl.software/FSL-1.1-ALv2.template.md
servers:
  - url: http://localhost:4000
    description: Local development
paths:
...
  /v1/playground/runs:
    get:
      operationId: listPlaygroundRuns
      summary: List recent Playground run history
      description: >
        Returns the customer's most-recent playground runs in reverse
        chronological order. Populated only when /v1/detect is called
        with source=playground (the dashboard does this automatically).
        Deliberately separate from traces — playground activity does
        not pollute production observability.
      tags: [Playground]
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
      responses:
        "200":
          description: Recent playground runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlaygroundRun"
  /v1/playground/runs/{id}:
    delete:
      operationId: deletePlaygroundRun
      summary: Delete a single playground run
      tags: [Playground]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # Custom Policies (tenant-owned overlays layered on top of the core security profile).
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "Bearer sk-bastio-..."
  schemas:
...
    DetectFinding:
      type: object
      properties:
        threat_type:
          type: string
          enum: [injection, pii, jailbreak]
        detector_name:
          type: string
        severity:
          type: string
          enum: [critical, high, medium, low, info]
        score:
          type: number
        confidence:
          type: number
        matched_pattern:
          type: string
        matched_content:
          type: string
        action:
          type: string
        message:
          type: string
    DetectStepResult:
      type: object
      properties:
        detector:
          type: string
        strategy:
          type: string
        fired:
          type: boolean
        skipped:
          type: boolean
          description: "True when the step was configured but not executed because an earlier step short-circuited the pipeline (e.g. block)."
        action:
          type: string
        score:
          type: number
        threshold:
          type: number
          description: "The weighted-score threshold this step was compared against. Useful for diagnosing fired=false at a high-looking score."
        findings:
          type: array
          items:
            $ref: "#/components/schemas/DetectFinding"
        duration:
          type: integer
          description: "Step duration in nanoseconds."
...
    PlaygroundRun:
      type: object
      properties:
        id:
          type: string
          format: uuid
        profile_name:
          type: string
        proxy_id:
          type: string
          format: uuid
          nullable: true
        direction:
          type: string
          enum: [input, output]
        prompt:
          type: string
        sanitized_content:
          type: string
        action:
          type: string
        should_block:
          type: boolean
        fired_detectors:
          type: array
          items:
            type: string
        steps:
          type: array
          items:
            $ref: "#/components/schemas/DetectStepResult"
        duration_ns:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time

    # ---- Custom Policies (tenant-owned overlays) ----
...